+++ webserver/webserver/src/main/java/io/helidon/webserver/UdsTransportBindingFactory.java
@@ -1,0 +1,75 @@
+/*
+ * Copyright (c) 2026 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.helidon.webserver;
+
+import java.util.Objects;
+
+import io.helidon.webserver.spi.TransportBinding;
+import io.helidon.webserver.spi.TransportBindingFactory;
+
+/**
+ * Unix domain socket transport binding factory.
+ */
+public final class UdsTransportBindingFactory implements TransportBindingFactory {
+    private final UdsTransportConfig config;
+
+    private UdsTransportBindingFactory(UdsTransportConfig config) {
+        this.config = Objects.requireNonNull(config, "config");
+    }
+
+    /**
+     * Create a Unix domain socket transport binding factory.
+     *
+     * @param config Unix domain socket transport binding configuration
+     * @return Unix domain socket transport binding factory
+     */
+    public static UdsTransportBindingFactory create(UdsTransportConfig config) {
+        return new UdsTransportBindingFactory(config);
+    }
+
+    @Override
+    public String type() {
+        return UdsTransportBinding.TYPE;
+    }
+
+    @Override
+    public String name() {
+        return config.name();
+    }
+
+    @Override
+    public boolean enabled() {
+        return config.enabled();
+    }
+
+    @Override
+    public boolean required() {
+        return config.required();
+    }
+
+    @Override
+    public boolean canBind(BindingPlanContext context) {
+        Objects.requireNonNull(context, "context");
+        return config.enabled() && config.socket().isPresent();
+    }
+
+    @Override
+    public TransportBinding create(TransportBindingContext context) {
+        Objects.requireNonNull(context, "context");
+        return new UdsTransportBinding(context, config);
+    }
+}
